fixing export and transport IDs
Here we remove the existing, automatically generated export reactions. We then re-add them so they match the new metabolite IDs and at the same time rename the transport reactions so that they all start with "T_".
Also the keys in the dictionary for cellular compartments has been changed.
N.B: again a lot changes in one go, making it hard to see.