Adding unique reactions
Here I added the unique reactions that were present in Matteo's version of the model. We see they don't cause huge discrepancies in the biomass prediction and so have been added to the working model here.

N.B: I reversed the previous commit, because something strange happened with the compartments in the model, corrupting the file. Here I fixed it and re-commit.