Adding ABC or PTS sugar transport rcts
In this commit, I used information from Brumm et al. (2015) and Beata's PhD thesis to add the ABC or PTS transport of sugars where there is a confirmation of growth on the substrate and bioinformatic evidence for the transport type.
Note: Gluconate, starch, xylan and pullulan are still left open to discuss in a meeting first. I will get back to those later.