fix rct
Here I fixed the BETALDHx and BETALDHy reactions as i saw the SBML validator was giving me an error. Looking back into my notebook 37 I saw a typo that added the reactants double to the BETALDH reactions and not to the BETALDHy reaction. So i fix that here and hope it solved the issue.