fix Biotin biosynthesis
In this commit, I've fixed biotin biosynthesis and added the biotin requirement to the biomass reaction.

Note: here, two demand reactions were necessary as we don't know what happens to the side products of these reactions. But they only use very small amounts of carbon flux so this is oke.