Adding Kegg gene ID to the genes in the model
Here I added the Kegg gene IDs in the annotation to the genes in our model. This was done based on using the e.c. code for the reactions associated to each gene. See notebook 53 for more info on how this is done.
N.B.: many of the changes show strangely in the changes here because so many things changed.